fix bug when catching up with live stream
when playing tracks from the currently on-air show, the last item_id is
not stable, so lookup can fail when the tracklist is re-fetched again
after the cache expires. to be exact, the last track in a stream has no
'offsetende' specified, so the id will change from '<startoffset>' to
'<startoffset>-<endoffset>'.